+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-page-views.js
@@ -56,7 +56,8 @@ window.GOVUK.analyticsGa4.analyticsModules = window.GOVUK.analyticsGa4.analytics
             dynamic: referrer ? 'true' : 'false',
             emergency_banner: document.querySelector('[data-ga4-emergency-banner]') ? 'true' : undefined,
             phase_banner: this.getElementAttribute('data-ga4-phase-banner') || undefined,
-            devolved_nations_banner: this.getElementAttribute('data-ga4-devolved-nations-banner') || undefined
+            devolved_nations_banner: this.getElementAttribute('data-ga4-devolved-nations-banner') || undefined,
+            cookie_banner: document.querySelector('[data-ga4-cookie-banner]') ? 'true' : undefined
           }
         }
         window.GOVUK.analyticsGa4.core.sendData(data)